requirements.txt
Flask==3.0.3
Pillow==10.4.0
openai==1.41.0
